Pawel's comment

+++ jupyter/experiment.py
@@ -56,7 +56,7 @@ class Experiment(object):
             Keys needed are: `ca_certs`, `keyfile`, `certfile` which are absolute paths,
                              `username` with `password` are mandatory,
                              `ssl_version` is optional and created in case of missing
-        :param read_csv: if no specify cassandra_cluster and port, try to read from a csv
+        :param read_csv: try to read  an experiment from from a csv in `dir_csv`
         :param dir_csv: save csv experiment to data directory for offline usage
 
         Initializes an experiment from a given experiment id by using the cassandra cluster and
@@ -66,47 +66,40 @@ class Experiment(object):
         if not os.path.exists(dir_csv):
             os.makedirs(dir_csv)
 
-        self.experiment_id = experiment_id
-        self.cassandra_cluster = cassandra_cluster
-        self.port = port
-        self.name = name if name else self.experiment_id
-        self.keyspace = keyspace
-        self.aggressor_throughput_namespaces_prefixes = aggressor_throughput_namespaces_prefixes
-        self.ssl_options = ssl_options
-        self.throughputs = defaultdict(list)  # keep throughputs from all aggressors to join it later with main DF
+        self.name = name if name else experiment_id
+        self.cached_experiment = os.path.join(dir_csv, '%s.csv' % experiment_id)
 
-        self.cached_experiment = os.path.join(dir_csv, '%s.csv' % self.experiment_id)
-
-        rows, qps = self.match_qps(read_csv)
-        self.frame = self.populate_data(rows, qps, read_csv)
-
-    def match_qps(self, read_csv):
         if read_csv:
-            return None, None
+            self.frame = pd.read_csv(self.cached_experiment)
         else:
-            session = Experiment._create_or_get_session(self.cassandra_cluster, self.port,
-                                                        self.ssl_options, self.keyspace)
-            rows = {}  # keep temporary rows from query for later match with qps rows
-            qps = {}  # qps is a one row from query, where we can map it to percentile rows
-            query = """SELECT ns, ver, host, time, boolval, doubleval, strval, tags, valtype
-                FROM snap.metrics WHERE tags['swan_experiment'] = \'%s\'""" % self.experiment_id
-            statement = SimpleStatement(query, fetch_size=100)
+            session_exec = self.read_data_from_cassandra(cassandra_cluster, port, keyspace, ssl_options, experiment_id)
+            rows, qps, throughputs = self.match_qps(session_exec, aggressor_throughput_namespaces_prefixes)
+            self.frame = self.create_df(rows, qps, throughputs)
 
-            for row in session.execute(statement):
-                k = (row.tags['swan_aggressor_name'], row.tags['swan_phase'], row.tags['swan_repetition'])
-                if row.ns == "/intel/swan/%s/%s/qps" % (row.ns.split("/")[3], row.host):
-                    qps[k] = row.doubleval
-                elif filter(lambda ns: ns in row.ns, self.aggressor_throughput_namespaces_prefixes):
-                    self.throughputs[k].append(row.doubleval)
-                else:
-                    rows[(row.ns,) + k] = row
-            return rows, qps
+    def match_qps(self, exec_session, aggressor_throughput_namespaces_prefixes):
+        rows = {}  # keep temporary rows from query for later match with qps rows
+        qps = {}  # qps is a one row from query, where we can map it to percentile rows
+        throughputs = defaultdict(list)  # keep throughputs from all aggressors to join it later with main DF
 
-    def populate_data(self, rows, qps, read_csv):
-        if read_csv:
-            frame = pd.read_csv(self.cached_experiment)
-            return frame
+        for row in exec_session:
+            k = (row.tags['swan_aggressor_name'], row.tags['swan_phase'], row.tags['swan_repetition'])
+            if row.ns == "/intel/swan/%s/%s/qps" % (row.ns.split("/")[3], row.host):
+                qps[k] = row.doubleval
+            elif filter(lambda ns: ns in row.ns, aggressor_throughput_namespaces_prefixes):
+                throughputs[k].append(row.doubleval)
+            else:
+                rows[(row.ns,) + k] = row
+        return rows, qps, throughputs
 
+    def read_data_from_cassandra(self, cassandra_cluster, port, keyspace, ssl_options, experiment_id):
+        session = Experiment._create_or_get_session(cassandra_cluster, port, keyspace, ssl_options)
+        query = """SELECT ns, ver, host, time, boolval, doubleval, strval, tags, valtype
+                        FROM snap.metrics WHERE tags['swan_experiment'] = \'%s\'""" % experiment_id
+        statement = SimpleStatement(query, fetch_size=100)
+
+        return session.execute(statement)
+
+    def create_df(self, rows, qps, throughputs):
         data = []
         for row in rows.itervalues():
             if row.valtype == "doubleval":
@@ -115,7 +108,7 @@ class Experiment(object):
                 achieved_qps = (qps[k] / float(row.tags['swan_loadpoint_qps']))
                 percent_qps = '{percent:.2%}'.format(percent=achieved_qps)
 
-                max_throughput = max(self.throughputs[k]) if k in self.throughputs else np.nan
+                max_throughput = max(throughputs[k]) if k in throughputs else np.nan
 
                 values = [row.ns, row.host, row.time, row.doubleval, row.tags['plugin_running_on'],
                           row.tags['swan_loadpoint_qps'], percent_qps, row.tags['swan_experiment'],
@@ -129,7 +122,8 @@ class Experiment(object):
                    'swan_repetition', 'throughputs']
         frame = pd.DataFrame(data, columns=columns)
 
-        frame.to_csv(self.cached_experiment)
+        if not os.path.exists(self.cached_experiment):
+            frame.to_csv(self.cached_experiment)
 
         return frame
